move analyze run on deposit button
